+++ src/pdf/base/XojCairoPdfExport.cpp
@@ -107,7 +107,7 @@ bool XojCairoPdfExport::createPdf(path file, PageRangeVector& range)
 		for (int i = e->getFirst(); i <= e->getLast(); i++)
 		{
 			int p = i - 1;
-			if (p < 0 || p >= doc->getPageCount())
+			if (p < 0 || p >= (int)doc->getPageCount())
 			{
 				continue;
 			}
